renamed setup.h -> struct.h for clarity
